Extract seed for EDM
