+++ lib/csdl2openapi.js
@@ -4,6 +4,13 @@
  * Latest version: https://github.com/oasis-tcs/odata-openapi/blob/master/lib/csdl2openapi.js
  */
 
+const {
+  EDM,
+  isIdentifier,
+  nameParts,
+  namespaceQualifiedName,
+} = require("./edm");
+
 //TODO
 // - Core.Example for complex types
 // - reduce number of loops over schemas
@@ -57,27 +64,20 @@ module.exports.csdl2openapi = function (
   const serviceRoot = scheme + "://" + host + basePath;
   const queryOptionPrefix = csdl.$Version <= "4.0" ? "$" : "";
   const typesToInline = {}; // filled in schema() and used in inlineTypes()
-  const boundOverloads = {};
-  const derivedTypes = {};
-  const alias = {};
-  const namespace = { Edm: "Edm" };
-  const namespaceUrl = {};
-  const voc = {};
-  const requiredSchemas = { list: [], used: {} };
   const oas31 = openapiVersion > "3.0.3";
+  const requiredSchemas = { list: [], used: {} };
 
-  preProcess(
-    csdl,
-    boundOverloads,
-    derivedTypes,
-    alias,
-    namespace,
-    namespaceUrl,
-    voc
-  );
+  const model = new EDM();
+  model.addDocument(csdl, messages);
+  //TODO: use directly from model, ideally remove
+  const boundOverloads = model.boundOverloads;
+  const derivedTypes = model.derivedTypes;
+  const namespace = model.namespace;
+  const namespaceUrl = model.namespaceUrl;
+  const voc = model.voc;
 
   const entityContainer = csdl.$EntityContainer
-    ? modelElement(csdl.$EntityContainer)
+    ? model.element(csdl.$EntityContainer)
     : {};
   const keyAsSegment = entityContainer[voc.Capabilities.KeyAsSegmentSupported];
   const applySupported = entityContainer[voc.Aggregation.ApplySupported];
@@ -101,202 +101,6 @@ module.exports.csdl2openapi = function (
   security(openapi, entityContainer);
 
   return openapi;
-
-  /**
-   * Collect model info for easier lookup
-   * @param {object} csdl CSDL document
-   * @param {object} boundOverloads Map of action/function names to bound overloads
-   * @param {object} derivedTypes Map of type names to derived types
-   * @param {object} alias Map of namespace or alias to alias
-   * @param {object} namespace Map of namespace or alias to namespace
-   * @param {object} namespaceUrl Map of namespace to reference URL
-   * @param {object} voc Map of vocabularies and terms
-   */
-  function preProcess(
-    csdl,
-    boundOverloads,
-    derivedTypes,
-    alias,
-    namespace,
-    namespaceUrl,
-    voc
-  ) {
-    Object.keys(csdl.$Reference || {}).forEach((url) => {
-      const reference = csdl.$Reference[url];
-      (reference.$Include || []).forEach((include) => {
-        const qualifier = include.$Alias || include.$Namespace;
-        alias[include.$Namespace] = qualifier;
-        namespace[qualifier] = include.$Namespace;
-        namespace[include.$Namespace] = include.$Namespace;
-        namespaceUrl[include.$Namespace] = url;
-      });
-    });
-
-    vocabularies(voc, alias);
-
-    Object.keys(csdl)
-      .filter((name) => isIdentifier(name))
-      .forEach((name) => {
-        const schema = csdl[name];
-        const qualifier = schema.$Alias || name;
-        const isDefaultNamespace = schema[voc.Core.DefaultNamespace];
-
-        alias[name] = qualifier;
-        namespace[qualifier] = name;
-        namespace[name] = name;
-
-        Object.keys(schema)
-          .filter((name) => isIdentifier(name))
-          .forEach((name) => {
-            const qualifiedName = qualifier + "." + name;
-            const element = schema[name];
-            if (Array.isArray(element)) {
-              element
-                .filter((overload) => overload.$IsBound)
-                .forEach((overload) => {
-                  const type =
-                    overload.$Parameter[0].$Type +
-                    (overload.$Parameter[0].$Collection ? "-c" : "");
-                  if (!boundOverloads[type]) boundOverloads[type] = [];
-                  boundOverloads[type].push({
-                    name: isDefaultNamespace ? name : qualifiedName,
-                    overload: overload,
-                  });
-                });
-            } else if (element.$BaseType) {
-              const base = namespaceQualifiedName(element.$BaseType);
-              if (!derivedTypes[base]) derivedTypes[base] = [];
-              derivedTypes[base].push(qualifiedName);
-            }
-          });
-
-        Object.keys(schema.$Annotations || {}).forEach((target) => {
-          const annotations = schema.$Annotations[target];
-          const segments = target.split("/");
-          const open = segments[0].indexOf("(");
-          let element;
-          if (open == -1) {
-            element = modelElement(segments[0]);
-          } else {
-            element = modelElement(segments[0].substring(0, open));
-            let args = segments[0].substring(open + 1, segments[0].length - 1);
-            element = element.find(
-              (overload) =>
-                (overload.$Kind == "Action" &&
-                  overload.$IsBound != true &&
-                  args == "") ||
-                (overload.$Kind == "Action" &&
-                  args ==
-                    (overload.$Parameter?.[0].$Collection
-                      ? `Collection(${overload.$Parameter[0].$Type})`
-                      : overload.$Parameter[0].$Type || "")) ||
-                (overload.$Parameter || [])
-                  .map((p) => {
-                    const type = p.$Type || "Edm.String";
-                    return p.$Collection ? `Collection(${type})` : type;
-                  })
-                  .join(",") == args
-            );
-          }
-          if (!element) {
-            messages.push(`Invalid annotation target '${target}'`);
-          } else if (Array.isArray(element)) {
-            //TODO: action or function:
-            //- loop over all overloads
-            //- if there are more segments, a parameter or the return type is targeted
-          } else {
-            switch (segments.length) {
-              case 1:
-                Object.assign(element, annotations);
-                break;
-              case 2:
-                if (["Action", "Function"].includes(element.$Kind)) {
-                  if (segments[1] == "$ReturnType") {
-                    if (element.$ReturnType)
-                      Object.assign(element.$ReturnType, annotations);
-                  } else {
-                    const parameter = element.$Parameter.find(
-                      (p) => p.$Name == segments[1]
-                    );
-                    Object.assign(parameter, annotations);
-                  }
-                } else {
-                  if (element[segments[1]]) {
-                    Object.assign(element[segments[1]], annotations);
-                  } else {
-                    messages.push(`Invalid annotation target '${target}'`);
-                  }
-                }
-                break;
-              default:
-                messages.push("More than two annotation target path segments");
-            }
-          }
-        });
-      });
-  }
-
-  /**
-   * Construct map of qualified term names
-   * @param {object} voc Map of vocabularies and terms
-   * @param {object} alias Map of namespace or alias to alias
-   */
-  function vocabularies(voc, alias) {
-    const terms = {
-      Aggregation: ["ApplySupported"],
-      Authorization: ["Authorizations", "SecuritySchemes"],
-      Capabilities: [
-        "BatchSupport",
-        "BatchSupported",
-        "ChangeTracking",
-        "CountRestrictions",
-        "DeleteRestrictions",
-        "DeepUpdateSupport",
-        "ExpandRestrictions",
-        "FilterRestrictions",
-        "IndexableByKey",
-        "InsertRestrictions",
-        "KeyAsSegmentSupported",
-        "NavigationRestrictions",
-        "OperationRestrictions",
-        "ReadRestrictions",
-        "SearchRestrictions",
-        "SelectSupport",
-        "SkipSupported",
-        "SortRestrictions",
-        "TopSupported",
-        "UpdateRestrictions",
-      ],
-      Core: [
-        "AcceptableMediaTypes",
-        "Computed",
-        "DefaultNamespace",
-        "Description",
-        "Example",
-        "Immutable",
-        "LongDescription",
-        "OptionalParameter",
-        "Permissions",
-        "SchemaVersion",
-      ],
-      JSON: ["Schema"],
-      Validation: [
-        "AllowedValues",
-        "Exclusive",
-        "Maximum",
-        "Minimum",
-        "Pattern",
-      ],
-    };
-
-    Object.keys(terms).forEach((vocab) => {
-      voc[vocab] = {};
-      const namespace = `Org.OData.${vocab}.V1`;
-      terms[vocab].forEach((term) => {
-        voc[vocab][term] = `@${alias[namespace] || namespace}.${term}`;
-      });
-    });
-  }
 
   /**
    * Construct the Info Object
@@ -385,7 +189,7 @@ module.exports.csdl2openapi = function (
                   property.$Kind == "NavigationProperty" ||
                   targetNP.qualifier != "Edm"
                 ) {
-                  const target = modelElement(property.$Type);
+                  const target = model.element(property.$Type);
                   const bidirectional =
                     property.$Partner &&
                     target &&
@@ -444,13 +248,13 @@ module.exports.csdl2openapi = function (
         } else {
           if (resource.$Action) {
             diagram += comma + "[" + name + color.resource + "]";
-            const overload = modelElement(resource.$Action).find(
-              (o) => !o.$IsBound
-            );
+            const overload = model
+              .element(resource.$Action)
+              .find((o) => !o.$IsBound);
             diagram += overloadDiagram(name, overload);
           } else if (resource.$Function) {
             diagram += comma + "[" + name + color.resource + "]";
-            const overloads = modelElement(resource.$Function);
+            const overloads = model.element(resource.$Function);
             if (overloads) {
               const unbound = overloads.filter((o) => !o.$IsBound);
               // TODO: loop over all overloads, add new source box after first arrow
@@ -494,7 +298,7 @@ module.exports.csdl2openapi = function (
     function overloadDiagram(name, overload) {
       let diag = "";
       if (overload.$ReturnType) {
-        const type = modelElement(overload.$ReturnType.$Type || "Edm.String");
+        const type = model.element(overload.$ReturnType.$Type || "Edm.String");
         if (type) {
           diag +=
             "-" +
@@ -505,7 +309,7 @@ module.exports.csdl2openapi = function (
         }
       }
       for (const param of overload.$Parameter || []) {
-        const type = modelElement(param.$Type || "Edm.String");
+        const type = model.element(param.$Type || "Edm.String");
         if (type) {
           diag +=
             comma +
@@ -521,17 +325,6 @@ module.exports.csdl2openapi = function (
       }
       return diag;
     }
-  }
-
-  /**
-   * Find model element by qualified name
-   * @param {string} qname Qualified name of model element
-   * @return {object} Model element
-   */
-  function modelElement(qname) {
-    const q = nameParts(qname);
-    const schema = csdl[q.qualifier] || csdl[namespace[q.qualifier]];
-    return schema ? schema[q.name] : null;
   }
 
   /**
@@ -556,7 +349,7 @@ module.exports.csdl2openapi = function (
       .filter((name) => isIdentifier(name) && container[name].$Type)
       .forEach((child) => {
         const tag = { name: child };
-        const type = modelElement(container[child].$Type);
+        const type = model.element(container[child].$Type);
         const description =
           container[child][voc.Core.Description] ||
           (type && type[voc.Core.Description]);
@@ -631,7 +424,7 @@ module.exports.csdl2openapi = function (
     navigationPath
   ) {
     const name = prefix.substring(prefix.lastIndexOf("/") + 1);
-    const type = modelElement(element.$Type);
+    const type = model.element(element.$Type);
     const pathItem = {};
     const restrictions = navigationPropertyRestrictions(root, navigationPath);
     const nonExpandable = nonExpandableProperties(root, navigationPath);
@@ -823,7 +616,7 @@ module.exports.csdl2openapi = function (
       (restrictions.IndexableByKey != false && targetIndexable)
     ) {
       const name = prefix.substring(prefix.lastIndexOf("/") + 1);
-      const type = modelElement(element.$Type);
+      const type = model.element(element.$Type);
       const key = entityKey(type, level);
       if (key.parameters.length > 0) {
         const path = prefix + key.segment;
@@ -923,7 +716,7 @@ module.exports.csdl2openapi = function (
       {};
 
     if (insertRestrictions.Insertable !== false) {
-      const type = modelElement(element.$Type);
+      const type = model.element(element.$Type);
       const hasStream = type && type.$HasStream;
       pathItem.post = {
         summary:
@@ -1197,7 +990,7 @@ module.exports.csdl2openapi = function (
    */
   function navigationPaths(element, prefix = "", level = 0) {
     const paths = [];
-    const type = modelElement(element.$Type);
+    const type = model.element(element.$Type);
     const properties = propertiesOfStructuredType(type);
     Object.keys(properties).forEach((key) => {
       if (properties[key].$Kind == "NavigationProperty") {
@@ -1307,7 +1100,7 @@ module.exports.csdl2openapi = function (
    */
   function primitivePaths(element, prefix = "") {
     const paths = [];
-    const elementType = modelElement(element.$Type);
+    const elementType = model.element(element.$Type);
 
     if (!elementType) {
       messages.push(`Unknown type for element: ${JSON.stringify(element)}`);
@@ -1319,7 +1112,7 @@ module.exports.csdl2openapi = function (
       .filter((entry) => entry[1].$Kind !== "NavigationProperty")
       .filter((entry) => entry[1].$Type)
       .filter((entry) => nameParts(entry[1].$Type).qualifier !== "Edm")
-      .filter((entry) => !modelElement(entry[1].$Type));
+      .filter((entry) => !model.element(entry[1].$Type));
 
     // Keep old logging
     ignore.forEach((entry) =>
@@ -1363,7 +1156,7 @@ module.exports.csdl2openapi = function (
     return function (entry) {
       const key = entry[0];
       const property = entry[1];
-      const propertyType = property.$Type && modelElement(property.$Type);
+      const propertyType = property.$Type && model.element(property.$Type);
 
       if (
         propertyType &&
@@ -1427,7 +1220,7 @@ module.exports.csdl2openapi = function (
       {};
 
     if (selectSupport.Supported !== false) {
-      const type = modelElement(element.$Type) || {};
+      const type = model.element(element.$Type) || {};
       const properties = propertiesOfStructuredType(type);
       const selectItems = [];
       Object.keys(properties)
@@ -1523,7 +1316,7 @@ module.exports.csdl2openapi = function (
       {};
 
     if (updateRestrictions.Updatable !== false) {
-      const type = modelElement(element.$Type);
+      const type = model.element(element.$Type);
       const operation = {
         summary:
           updateRestrictions.Description ||
@@ -1740,7 +1533,7 @@ module.exports.csdl2openapi = function (
         level < maxLevels
       ) {
         navigationPathMap(
-          modelElement(properties[key].$Type),
+          model.element(properties[key].$Type),
           map,
           prefix + key + "/",
           level + 1
@@ -1777,7 +1570,7 @@ module.exports.csdl2openapi = function (
     while (type) {
       keys = type.$Key;
       if (keys || !type.$BaseType) break;
-      type = modelElement(type.$BaseType);
+      type = model.element(type.$BaseType);
     }
     return keys;
   }
@@ -1811,7 +1604,7 @@ module.exports.csdl2openapi = function (
         const segments = key[parameter].split("/");
         property = properties[segments[0]];
         for (let i = 1; i < segments.length; i++) {
-          const complexType = modelElement(property.$Type);
+          const complexType = model.element(property.$Type);
           const properties = propertiesOfStructuredType(complexType);
           property = properties[segments[i]];
         }
@@ -1952,9 +1745,9 @@ module.exports.csdl2openapi = function (
    * @param {object} child Action import object
    */
   function pathItemActionImport(paths, name, child) {
-    const overload = modelElement(child.$Action).find(
-      (overload) => !overload.$IsBound
-    );
+    const overload = model
+      .element(child.$Action)
+      .find((overload) => !overload.$IsBound);
     pathItemAction(
       paths,
       "/" + name,
@@ -2048,7 +1841,7 @@ module.exports.csdl2openapi = function (
    * @param {object} child Function import object
    */
   function pathItemFunctionImport(paths, name, child) {
-    const overloads = modelElement(child.$Function);
+    const overloads = model.element(child.$Function);
     overloads &&
       overloads
         .filter((overload) => !overload.$IsBound)
@@ -2107,7 +1900,7 @@ module.exports.csdl2openapi = function (
         .filter((t) => t)
         .join("  \n");
       if (description) param.description = description;
-      const type = modelElement(p.$Type || "Edm.String");
+      const type = model.element(p.$Type || "Edm.String");
       if (
         p.$Collection ||
         p.$Type == "Edm.Stream" ||
@@ -2131,7 +1924,7 @@ module.exports.csdl2openapi = function (
           "URL-encoded JSON " +
           (p.$Collection ? "array with items " : "") +
           "of type " +
-          namespaceQualifiedName(p.$Type || "Edm.String") +
+          namespaceQualifiedName(namespace, p.$Type || "Edm.String") +
           ", see [Complex and Collection Literals](https://docs.oasis-open.org/odata/odata/v4.01/odata-v4.01-part2-url-conventions.html#sec_ComplexandCollectionLiterals)";
         param.example = p.$Collection ? "[]" : "{}";
       } else {
@@ -2246,7 +2039,7 @@ module.exports.csdl2openapi = function (
     }
 
     if (updateRestrictions.FilterSegmentSupported) {
-      const type = modelElement(element.$Type);
+      const type = model.element(element.$Type);
       pathItem.patch = {
         summary:
           "Update entities matching the filter expression in " +
@@ -2468,7 +2261,7 @@ module.exports.csdl2openapi = function (
     const unordered = {};
 
     for (const r of requiredSchemas.list) {
-      const type = modelElement(`${r.namespace}.${r.name}`);
+      const type = model.element(`${r.namespace}.${r.name}`);
       if (!type) continue;
       switch (type.$Kind) {
         case "ComplexType":
@@ -2686,7 +2479,7 @@ module.exports.csdl2openapi = function (
   function propertiesOfStructuredType(type) {
     const properties =
       type && type.$BaseType
-        ? propertiesOfStructuredType(modelElement(type.$BaseType))
+        ? propertiesOfStructuredType(model.element(type.$BaseType))
         : {};
     if (type) {
       Object.keys(type)
@@ -3054,7 +2847,7 @@ module.exports.csdl2openapi = function (
         if (element.$Type.startsWith("Edm.")) {
           messages.push("Unknown type: " + element.$Type);
         } else {
-          let type = modelElement(element.$Type);
+          let type = model.element(element.$Type);
           let isStructured =
             type && ["ComplexType", "EntityType"].includes(type.$Kind);
           s = ref(element.$Type, isStructured ? suffix : "");
@@ -3305,37 +3098,5 @@ module.exports.csdl2openapi = function (
       s[scheme.Authorization] = scheme.RequiredScopes || [];
       openapi.security.push(s);
     });
-  }
-
-  /**
-   * a qualified name consists of a namespace or alias, a dot, and a simple name
-   * @param {string} qualifiedName
-   * @return {string} namespace-qualified name
-   */
-  function namespaceQualifiedName(qualifiedName) {
-    let np = nameParts(qualifiedName);
-    return namespace[np.qualifier] + "." + np.name;
-  }
-
-  /**
-   * a qualified name consists of a namespace or alias, a dot, and a simple name
-   * @param {string} qualifiedName
-   * @return {object} with components qualifier and name
-   */
-  function nameParts(qualifiedName) {
-    const pos = qualifiedName.lastIndexOf(".");
-    return {
-      qualifier: qualifiedName.substring(0, pos),
-      name: qualifiedName.substring(pos + 1),
-    };
-  }
-
-  /**
-   * an identifier does not start with $ and does not contain @
-   * @param {string} name
-   * @return {boolean} name is an identifier
-   */
-  function isIdentifier(name) {
-    return !name.startsWith("$") && !name.includes("@");
   }
 };
